cacto-chan/requirements.txt
git+https://github.com/Rapptz/discord.py
dnspython==1.16.0
async-timeout==3.0.1
discord.py==1.3.4+
pycparser==2.20
PyNaCl==1.3.0
six==1.14.0
websockets==8.1
pandas
